requirements.txt
autopep8==1.7.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
google-auth==2.11.0
google-auth-oauthlib==0.5.2
gspread==5.5.0
idna==3.3
lxml==4.9.1
numpy==1.23.2
oauthlib==3.2.0
pandas==1.4.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
urllib3==1.26.12
